moved common part out of block_layout_comverter
