+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/block_layout_converter.h
@@ -6,20 +6,12 @@
 #include <arrow/type.h>
 
 #include <ydb/library/yql/dq/comp_nodes/hash_join_utils/tuple.h>
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/layout_converter_common.h>
 
 namespace NKikimr::NMiniKQL {
 
 class IBlockLayoutConverter : private TNonCopyable {
 public:
-    struct TPackResult {
-        std::vector<ui8, TMKQLAllocator<ui8>> PackedTuples;
-        std::vector<ui8, TMKQLAllocator<ui8>> Overflow;
-        int NTuples{0};
-    };
-
-    using TPackedTuple = std::vector<ui8, TMKQLAllocator<ui8>>;
-    using TOverflow = std::vector<ui8, TMKQLAllocator<ui8>>;
-
 public:
     using TPtr = std::unique_ptr<IBlockLayoutConverter>;
